Refactored the migration component
